Refactor challenge helpers and add pipeline smoke test
